requirements.txt
# Backend Dependencies for both GCS and RecordingAnalysis
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
boto3==1.34.0
python-multipart==0.0.6
pymavlink==2.4.41
opencv-contrib-python==4.12.0.88
ultralytics==8.4.3
torch==2.5.1
geographiclib==2.1
# Testing Dependencies
pytest
pytest-asyncio
pytest-cov
httpx>=0.27.0